Monolingual+Multilingual+Crosslingual experiments

+++ src/generation.py
@@ -47,15 +47,56 @@ def formatting_func_factory(tokenizer, args):
     system_message['nl'] = "Je bent een lexicograaf die vertrouwd is met het geven van beknopte definities van woordbetekenissen."
     system_message['it'] = "Sei un lessicografo esperto nel fornire definizioni concise dei significati delle parole."
     system_message['sv'] = "Du är en lexikograf som är van vid att ge kortfattade definitioner av ordens betydelser."
+    system_message['no'] = "Du er en leksikograf som er kjent med å gi presise definisjoner av ords betydning."
+    system_message['es'] = "Eres un lexicógrafo familiarizado con proporcionar definiciones concisas de los significados de las palabras."
+    system_message['ja'] = "あなたは、単語の意味の簡潔な定義を提供することに熟練した辞書編纂者です。"
+    system_message['de'] = "Du bist ein Lexikograf, der mit der Bereitstellung prägnanter Definitionen von Wortbedeutungen vertraut ist."
+    system_message['pt'] = "Você é um lexicógrafo familiarizado com a fornecimento de definições concisas dos significados das palavras."   
+    system_message['en'] = "You are a lexicographer familiar with providing concise definitions of word meanings."
+    system_message['tr'] = "Sen, kelime anlamlarının özlü tanımlarını sağlamaya aşina bir sözlük yazarsısın."
+    system_message['mg'] = "Ianao dia lexicographer mahazatra amin'ny fanomezana fanazavana fohy momba ny dikan'ny teny."
+    system_message['da'] = "Du er en leksikograf, der er vant til at give præcise definitioner af ords betydninger."
+    system_message['ca'] = "Ets un lexicògraf familiaritzat amb la creació de definicions concises dels significats de les paraules."
+    system_message['fr'] = "Vous êtes un lexicographe habitué à fournir des définitions concises des significations des mots."
+    system_message['lt'] = "Jūs esate leksikografas, kuris gerai susipažinęs su trumpų žodžių reikšmių apibrėžimų pateikimu."
+    system_message['la'] = "Es lexicographus peritus, qui breves definitiones significatuum verborum praebet."
+    system_message['id'] = "Anda adalah seorang leksikograf yang terbiasa memberikan definisi singkat dari makna kata-kata."
+    system_message['pl'] = "Jesteś leksykografem, który zna się na podawaniu zwięzłych definicji znaczeń słów."
+    system_message['ku'] = "Hûn lexicographer in ku bi dayîna şîroveyên kurt ên maneya peyvên nasnamekî ne."
+    system_message['el'] = "Είστε ένας λεξικογράφος εξοικειωμένος με την παροχή συνοπτικών ορισμών των εννοιών των λέξεων."
+    system_message['zh'] = "你是一位熟悉提供简明单词含义定义的词典编纂者。"
+    system_message['fi'] = "Olet sanakirjantekijä, joka tuntee sanan merkitysten ytimekkäiden määritelmien antamisen."
+    system_message['ru'] = "Вы — лексикограф, знакомый с составлением кратких определений значений слов."
 
     user_message = dict()
     user_message['nl'] = 'Geef alstublieft een beknopte definitie van de betekenis van het woord "{}" in de volgende zin: {}'
     user_message['it'] = 'Si prega di fornire una definizione concisa per il significato della parola "{}" nella seguente frase: {}'
     user_message['sv'] = 'Vänligen ge en kortfattad definition av betydelsen av ordet "{}" i följande mening: {}'
-    
+    user_message['es'] = 'Por favor, proporcione una definición concisa para el significado de la palabra "{}" en la siguiente oración: {}'
+    user_message['no'] = 'Vennligst gi en kortfattet definisjon av betydningen av ordet "{}" i den følgende setningen: {}'
+    user_message['ja'] = '次の文での「{}」という単語の意味に対する簡潔な定義を提供してください: {}'
+    user_message['de'] = 'Bitte geben Sie eine prägnante Definition für die Bedeutung des Wortes "{}" im folgenden Satz an: {}'
+    user_message['pt'] = 'Por favor, forneça uma definição concisa para o significado da palavra "{}" na seguinte frase: {}'
+    user_message['en'] = 'Please provide a concise definition for the meaning of the word "{}" in the following sentence: {}'
+    user_message['tr'] = 'Lütfen aşağıdaki cümledeki "{}" kelimesinin anlamı için özlü bir tanım sağlayın: {}'
+    user_message['mg'] = 'Azafady, omeo fanazavana fohy momba ny dikan\'ny teny "{}" ao amin\'ity fehezanteny manaraka ity: {}'
+    user_message['da'] = 'Venligst giv en kortfattet definition af betydningen af ordet "{}" i den følgende sætning: {}'
+    user_message['ca'] = 'Si us plau, proporcioneu una definició concisa del significat de la paraula "{}" en la següent frase: {}'
+    user_message['fr'] = 'Veuillez fournir une définition concise du sens du mot "{}" dans la phrase suivante : {}'
+    user_message['lt'] = 'Prašome pateikti trumpą žodžio "{}" reikšmės apibrėžimą šioje sakinyje: {}'
+    user_message['la'] = 'Quaeso, praebe brevem definitionem significatuum verbi "{}" in sequenti sententia: {}'
+    user_message['id'] = 'Tolong berikan definisi singkat untuk makna kata "{}" dalam kalimat berikut: {}'
+    user_message['pl'] = 'Proszę podać zwięzłą definicję znaczenia słowa "{}" w następującym zdaniu: {}'
+    user_message['ku'] = 'Ji kerema xwe, daxuyaniya kurt ji bo maneya peyva "{}" di gotarê jêrîn de pêşkêş bikin: {}'
+    user_message['el'] = 'Παρακαλώ παρέχετε έναν συνοπτικό ορισμό για τη σημασία της λέξης "{}" στην παρακάτω πρόταση: {}'
+    user_message['zh'] = '请提供单词"{}"在以下句子中的简洁定义：{}'
+    user_message['fi'] = 'Ole hyvä ja anna lyhyt määritelmä sanan "{}" merkitykselle seuraavassa lauseessa: {}'
+    user_message['ru'] = 'Пожалуйста, предоставьте краткое определение значения слова "{}" в следующем предложении: {}'
+
     def formatting_func(record):
-        return tokenizer.apply_chat_template([{'role': 'system', 'content': system_message[args.language]},
-                                              {'role': 'user', 'content': user_message[args.language].format(record['target'], record['example'])}],
+        language = record['language'] # args.language
+        return tokenizer.apply_chat_template([{'role': 'system', 'content': system_message[language]},
+                                              {'role': 'user', 'content': user_message[language].format(record['target'], record['example'])}],
                                                tokenize=False, add_generation_prompt=True)
     return formatting_func
 
@@ -111,10 +152,15 @@ if __name__ == '__main__':
     pipe = pipeline("text-generation", model=model, tokenizer=tokenizer, device_map="auto")
 
     for filename in args.test_filename:
+        stem = pathlib.Path(filename).stem
+        if pathlib.Path(f'{args.output_dir}/{stem}.txt').exists():
+            print('File exists:', f'{args.output_dir}/{stem}.txt')
+            continue
+        
         datasets = load_dataset('json', data_files=filename, split='train')
         outputs = generation(pipe, datasets, args)
 
-        stem = pathlib.Path(filename).stem
+        
         pathlib.Path(args.output_dir).mkdir(parents=True, exist_ok=True)
         with open(f'{args.output_dir}/{stem}.txt', mode='w', encoding='utf-8') as f:
             f.writelines(outputs)
